src/main/java/pl/mlopatka/messages/message/MessageController.java
package pl.mlopatka.messages.message;

import lombok.AllArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.lang.NonNull;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/messages")
@AllArgsConstructor
public class MessageController {

    private final MessageService messageService;

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public long createMessage(@NonNull @RequestBody MessageDto messageDto) {
        return messageService.createMessage(messageDto);
    }

    @GetMapping("/{author}")
    public List<MessageDto> getMessages(@NonNull @PathVariable("author") String author) {
        return messageService.getMessages(author);
    }
}
